// xref-helper/src/commands/types/ItemType.ts
/**
 * The search result `itemType`.
 */
export enum ItemType {
    namespace = 'Namespace',
    class = 'Class',
    struct = 'Struct',
    interface = 'Interface',
    enum = 'Enum',
    delegate = 'Delegate',
    method = 'Method',
    constructor = 'Constructor',
    operator = 'Operator',
    field = 'Field',
    member = 'Member',
    property = 'Property',
    event = 'Event',
    attachedProperty = 'AttachedProperty',
    attachedEvent = 'AttachedEvent',
    command = 'Command',
    module = 'Module',
    typeAlias = 'Type Alias',
    function = 'Function',
    rootImportPackage = 'RootImportPackage',
    importPackage = 'ImportPackage',
    package = 'Package',
}